Working analyzer
