+++ analyzer/package.json
@@ -1,8 +1,12 @@
 {
-  "name": "lona-extractor",
+  "name": "lona-analyzer",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
+  "scripts": {
+    "snapshot":
+      "node index.js > workspace/View1.component && node ../compiler/core/src/main.bs.js workspace swift workspace out"
+  },
   "dependencies": {
     "@babel/traverse": "^7.0.0-beta.44",
     "babel-core": "^6.26.0",
