+++ analyzer/workspace/View1.component
@@ -1,0 +1,29 @@
+{
+  "root": {
+    "id": "Test",
+    "type": "Lona:View",
+    "params": {
+      "flex": 1,
+      "justifyContent": "center",
+      "alignItems": "center"
+    },
+    "children": [
+      {
+        "id": "Test",
+        "type": "Lona:View",
+        "params": {
+          "width": 200,
+          "height": 200,
+          "backgroundColor": "skyblue",
+          "borderWidth": 2,
+          "borderColor": "steelblue",
+          "borderRadius": 20
+        },
+        "children": []
+      }
+    ]
+  },
+  "logic": [],
+  "params": [],
+  "canvases": []
+}
