+++ analyzer/workspace/colors.json
@@ -1,0 +1,3 @@
+{
+  "colors": []
+}
