+++ analyzer/workspace/textStyles.json
@@ -1,0 +1,3 @@
+{
+  "styles": []
+}
